a view fixes

+++ src/store/modules/authentication.js
@@ -1,4 +1,4 @@
-//import StreamsServices from "../../services/StreamsServices";
+import StreamsServices from "../../services/StreamsServices";
 import CookieService from "../../services/CookieSerice";
 
 export const namespaced = true;
@@ -21,50 +21,39 @@ export const mutations = {
 
 export const actions = {
     login({ commit }, userData) {
-        // return StreamsServices.postLogin(userData).then(response => {
-        //     commit("SET_ALL_UserData",{
-        //         success:response.data.success,
-        //         name:userData.username,
-        //         token:response.data.token
-        //     });
-        //     if(response.data.success)
-        //         CookieService.setToken(response.data.token)
-        //     return response.data.success;
-        // });
-
-        //TODO: remove fake Data
-        commit("SET_ALL_UserData", {
-            success: true,
-            name: userData.name,
-            token: "fake_token"
+        return StreamsServices.postLogin(userData).then(response => {
+            commit("SET_ALL_UserData", {
+                success: response.data.success,
+                name: userData.username,
+                token: response.data.token
+            });
+            if (response.data.success)
+                CookieService.setToken(response.data.token)
+            return response.data.success;
         });
-        CookieService.setToken("fake_token");
-        return true;
-
     },
-    authenticate({ commit }, token) {
-        // return StreamsServices.postToken(token).then(response => {
-        //     if (response.data.success) {
-        //         commit("SET_ALL_UserData", { token: token, success: true, name: response.username })
-        //     } else {
-        //         commit("SET_ALL_UserData", { token: null, loggedIn: false, name: null })
-        //         CookieService.removeToken();
-        //     }
-        //     return response.data.success;
-        // });
-
-        //TODO: remove fake Data
-        commit("SET_ALL_UserData", {
-            success: true,
-            name: "fake_user",
-            token: token
+    authenticate({ commit }) {
+        let token = CookieService.getToken();
+        if (!token) return false;
+        return StreamsServices.postToken(token).then(response => {
+            if (response.data.success) {
+                commit("SET_ALL_UserData", { token: token, success: true, name: response.username })
+            } else {
+                commit("SET_ALL_UserData", { token: null, loggedIn: false, name: null })
+                CookieService.removeToken();
+            }
+            return response.data.success;
         });
-        CookieService.setToken("fake_token");
-        return true;
-
     },
     logout({ commit }) {
-        commit("SET_ALL_UserData", { token: null, loggedIn: false, name: null })
-        CookieService.removeToken();
+        let token = CookieService.getToken();
+        if (!token) return false;
+        return StreamsServices.postLogout(token).then(response => {
+            if (response.data.success) {
+                commit("SET_ALL_UserData", { token: null, loggedIn: false, name: null })
+                CookieService.removeToken();
+            }
+            return response.data.success;
+        });
     }
 };
